Improve mongomock tests to avoid regression
